# openapi.json
{
  "components": {
    "schemas": {
...
      "FriendActionOut": {
        "description": "Result of a friend-graph mutation. The client re-fetches the lists after,\nso this only needs to say WHAT happened (used to phrase a toast).",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "FriendActionOut",
        "type": "object"
      },
      "FriendOut": {
        "description": "An accepted friend. `home` is populated ONLY when that friend has both set\na home AND turned sharing on — otherwise null (they stay listed, no marker).",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "home": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HomePointOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "online": {
            "default": false,
            "title": "Online",
            "type": "boolean"
          }
        },
        "required": [
          "id"
        ],
        "title": "FriendOut",
        "type": "object"
      },
      "FriendRequestOut": {
        "description": "One pending friend request, from the current user's point of view.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/FriendUserBrief"
          }
        },
        "required": [
          "id",
          "direction",
          "user",
          "created_at"
        ],
        "title": "FriendRequestOut",
        "type": "object"
      },
      "FriendRequestsOut": {
        "properties": {
          "incoming": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FriendRequestOut"
            },
            "title": "Incoming",
            "type": "array"
          },
          "outgoing": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FriendRequestOut"
            },
            "title": "Outgoing",
            "type": "array"
          }
        },
        "title": "FriendRequestsOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HomePointOut": {
        "description": "A friend's shared home coordinates — a map marker only (no radius).",
        "properties": {
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lon"
        ],
        "title": "HomePointOut",
        "type": "object"
      },
      "HomeShareIn": {
        "description": "PUT /me/home — set/update home coordinates and the share flag together.",
        "properties": {
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          },
          "share": {
            "default": true,
            "title": "Share",
            "type": "boolean"
          }
        },
        "required": [
          "lat",
          "lon"
        ],
        "title": "HomeShareIn",
        "type": "object"
      },
...
      "MyHomeOut": {
        "description": "GET /me/home — the current user's own stored home + share state.",
        "properties": {
          "home": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HomePointOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "share_home": {
            "title": "Share Home",
            "type": "boolean"
          }
        },
        "required": [
          "share_home"
        ],
        "title": "MyHomeOut",
        "type": "object"
      },
...
      "PresencePrefIn": {
        "description": "PUT /me/presence — opt in/out of showing friends WHEN you were last active.",
        "properties": {
          "share_presence": {
            "title": "Share Presence",
            "type": "boolean"
          }
        },
        "required": [
          "share_presence"
        ],
        "title": "PresencePrefIn",
        "type": "object"
      },
      "PresencePrefOut": {
        "properties": {
          "share_presence": {
            "title": "Share Presence",
            "type": "boolean"
          }
        },
        "required": [
          "share_presence"
        ],
        "title": "PresencePrefOut",
        "type": "object"
      },
...
      "SendFriendRequestIn": {
        "description": "POST /friends/requests — address a friend request by their email.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "SendFriendRequestIn",
        "type": "object"
      },
      "ShareToggleIn": {
        "description": "PATCH /me/home/share — flip sharing without touching coordinates.",
        "properties": {
          "share": {
            "title": "Share",
            "type": "boolean"
          }
        },
        "required": [
          "share"
        ],
        "title": "ShareToggleIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Kyiv Aerial Threat Tracker (MVP)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/friends": {
      "get": {
        "operationId": "list_friends_friends_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendOut"
                  },
                  "title": "Response List Friends Friends Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Friends",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/requests": {
      "get": {
        "operationId": "list_requests_friends_requests_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendRequestsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Requests",
        "tags": [
          "friends"
        ]
      },
      "post": {
        "operationId": "send_request_friends_requests_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendFriendRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Request",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/requests/{request_id}/accept": {
      "post": {
        "operationId": "accept_request_friends_requests__request_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Request",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/requests/{request_id}/decline": {
      "post": {
        "operationId": "decline_request_friends_requests__request_id__decline_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decline Request",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/{user_id}": {
      "delete": {
        "operationId": "remove_friend_friends__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Friend",
        "tags": [
          "friends"
        ]
      }
    },
...
    "/me/home": {
      "delete": {
        "operationId": "delete_my_home_me_home_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyHomeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete My Home",
        "tags": [
          "friends"
        ]
      },
      "get": {
        "operationId": "get_my_home_me_home_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyHomeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Home",
        "tags": [
          "friends"
        ]
      },
      "put": {
        "operationId": "put_my_home_me_home_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeShareIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyHomeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put My Home",
        "tags": [
          "friends"
        ]
      }
    },
    "/me/home/share": {
      "patch": {
        "operationId": "patch_home_share_me_home_share_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareToggleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyHomeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Home Share",
        "tags": [
          "friends"
        ]
      }
    },
    "/me/presence": {
      "get": {
        "operationId": "get_my_presence_me_presence_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresencePrefOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Presence",
        "tags": [
          "friends"
        ]
      },
      "put": {
        "description": "Opt in/out of disclosing WHEN you were last active. Turning it off hides\nthe timestamp immediately; it does not stop `last_seen_at` being recorded,\nsince the online dot is derived from the same field.",
        "operationId": "put_my_presence_me_presence_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresencePrefIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresencePrefOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put My Presence",
        "tags": [
          "friends"
        ]
      }
    },
...
  }
}